--- src/assets/styles/mixins/formField.scss
@@ -1,22 +1,0 @@
-@mixin form-field {
-  inline-size: 100%;
-  border: 2px solid transparent;
-  border-radius: 12px;
-  background:
-    linear-gradient(var(--clr-main), var(--clr-main)) padding-box,
-    linear-gradient(
-        180deg,
-        var(--clr-primary-input) 0%,
-        var(--clr-primary-default) 100%
-      )
-      border-box;
-  font-family: var(--ffm-default);
-  font-size: 16px;
-  padding: 12px 16px;
-  transition: border-color 0.3s;
-
-  &:focus {
-    outline: none;
-    border-color: var(--clr-blue);
-  }
-}
