--- src/shared/ui/formField/index.module.scss
@@ -1,36 +1,0 @@
-@use '../../../assets/styles/mixins/formField.scss' as *;
-
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  margin-bottom: 32px;
-  position: relative;
-
-  &.form--has-error {
-    & textarea,
-    & input {
-      border-color: var(--clr-secondary-default);
-    }
-  }
-}
-
-.label {
-  display: block;
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 1.2;
-}
-
-.error {
-  position: absolute;
-  bottom: -20px;
-  left: 0;
-  font-family: var(--ffm-default);
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 1.4;
-  color: var(--clr-secondary-default);
-  min-height: 20px;
-  margin: 0;
-}
