+++ src/shared/ui/textarea/index.module.scss
@@ -7,53 +7,51 @@
   gap: 16px;
   font-weight: 500;
   line-height: 1.36;
-position: relative;
+  position: relative;
 
-&__label {
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 1;
-}
-
-&__area {
-  display: flex;
-  align-items: flex-start;
-  padding: 12px 16px;
-  inline-size: 100%;
-  min-block-size: 164px;
-  font-family: var(--ffm-default);
-  font-size: 15px;
-  line-height: 1.6;
-  resize: vertical;
-  border-radius: var(--border-radius-btn-social);
-
-  @include inputs-gradient;
-  
-
-  &:focus {
-    outline: none;
+  &__label {
+    font-weight: 700;
+    font-size: 20px;
+    line-height: 1;
   }
 
-   &.error {
+  &__area {
+    display: flex;
+    align-items: flex-start;
+    padding: 12px 16px;
+    inline-size: 100%;
+    min-block-size: 164px;
+    font-family: var(--ffm-default);
+    font-size: 15px;
+    line-height: 1.6;
+    resize: vertical;
+    border-radius: var(--border-radius-btn-social);
+
+    @include inputs-gradient;
+
+    &:focus {
+      outline: none;
+    }
+
+    &.error {
       border: 2px solid var(--clr-secondary-default);
     }
 
-  &::placeholder {
-    color: var(--clr-4, #858585);
+    &::placeholder {
+      color: var(--clr-4, #858585);
+    }
+  }
+
+  &__error {
+    position: absolute;
+    bottom: -28px;
+    left: 0;
+    font-family: var(--ffm-default);
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 1.5;
+    color: var(--clr-secondary-default);
+    min-height: 18px;
+    margin: 0;
   }
 }
-
-&__error {
-  position: absolute;
-  bottom: -28px;
-  left: 0;
-  font-family: var(--ffm-default);
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 1.5;
-  color: var(--clr-secondary-default);
-  min-height: 18px;
-  margin: 0;
-}
-}
-
