+++ src/shared/ui/textarea/type.ts
@@ -1,0 +1,8 @@
+type TextareaProps = {
+  label?: string;
+  error?: string;
+  htmlFor?: string;
+  classname?: string;
+} & React.TextareaHTMLAttributes<HTMLTextAreaElement>;
+
+export default TextareaProps
